+++ src/infinity/facets/Bridges/Toucan/RedeemToucanPoolFacet.sol
@@ -75,8 +75,7 @@ contract RedeemToucanPoolFacet is ReentrancyGuard {
         require(toMode == LibTransfer.To.EXTERNAL, "Internal balances not live");
         require(projectTokens.length == amounts.length, "Array lengths not equal");
 
-
-        uint totalCarbon;
+        uint256 totalCarbon;
         address originalSourceToken = sourceToken;
 
         for (uint256 i; i < amounts.length; i++) {
